+++ src/components/FormPanel/FormPanel.test.tsx
@@ -1529,7 +1529,7 @@ describe('Panel', () => {
             options: {
               initial: {
                 method: RequestMethod.NONE,
-                code: 'notifySuccess("success"); notifyError("error"); notifyWarning("warning");',
+                code: 'context.grafana.notifySuccess("success"); context.grafana.notifyError("error"); context.grafana.notifyWarning("warning");',
               },
             },
           })
@@ -1537,7 +1537,7 @@ describe('Panel', () => {
       );
 
       expect(replaceVariables).toHaveBeenCalledWith(
-        'notifySuccess("success"); notifyError("error"); notifyWarning("warning");'
+        'context.grafana.notifySuccess("success"); context.grafana.notifyError("error"); context.grafana.notifyWarning("warning");'
       );
       expect(appEventsMock.publish).toHaveBeenCalledWith({
         type: AppEvents.alertSuccess.name,
@@ -1563,7 +1563,7 @@ describe('Panel', () => {
         initial: {},
         update: {
           method: RequestMethod.NONE,
-          code: 'notifySuccess("success"); notifyError("error");',
+          code: 'context.grafana.notifySuccess("success"); context.grafana.notifyError("error");',
         },
       };
       const { rerender } = await act(async () =>
@@ -1682,12 +1682,12 @@ describe('Panel', () => {
 
       const defaultOptions = {
         initial: {
-          code: `notifyError("error");`,
+          code: `context.grafana.notifyError("error");`,
         },
         update: {},
         resetAction: {
           mode: ResetActionMode.CUSTOM,
-          code: 'notifySuccess("success");',
+          code: 'context.grafana.notifySuccess("success");',
         },
       };
       const { rerender } = await act(async () =>
@@ -1919,12 +1919,12 @@ describe('Panel', () => {
 
       const defaultOptions = {
         initial: {
-          code: `notifySuccess("success");`,
+          code: `context.grafana.notifySuccess("success");`,
         },
         update: {},
         resetAction: {
           mode: ResetActionMode.CUSTOM,
-          code: 'notifySuccess("success");',
+          code: 'context.grafana.notifySuccess("success");',
           confirm: true,
         },
       };